+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - `WithDetails` now accepts optional pairs of keys and values as initial details.
+- `Errorf` supports multiple `%w`.
 
 ## [0.5.0] - 2023-06-06